Renamed ChatConnect as useChat
This was done to indicate that it is a hook (following React's strong
recommendation that hooks be named with the "use" prefix).

Also changed the return object of the hook, in order to expose the API
for sending messages and checking connectivity rather than returning the
entire socket object.
